Chore: Migrate TaskQueue.test to TypeScript
TaskQueue.ts types the constructor as requiring a `name` and `push`
as taking a `string` id, but the old JS test passed no name and
numeric ids. The behaviour was identical (JS object property
access coerces numbers to strings), so the test now passes a name
and uses string ids to match the typed signature.

Change by Claude Code
